CustomProgressIndicator 추가, tmp 폴더에 업로드한 파일 삭제 로직 추가, WidgetBinding.instance.addPostFrameCallback 추가

+++ lib/components/FileProcessing.dart
@@ -1,8 +1,6 @@
 import 'dart:convert';
-import 'dart:io';
 import 'dart:math';
 import 'dart:typed_data';
-
 import 'package:file_picker/file_picker.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:http/http.dart' as http;
@@ -12,6 +10,7 @@ import 'CustomClass/CustomToast.dart';
 
 class FileProcessing {
   static final FirebaseStorage _storageRef = FirebaseStorage.instance;
+  static final List<String> allowedExtensions = ['jpg', 'jpeg', 'png', 'pdf'];
 
   static Future<Uint8List?> loadFileFromStorage(String? relativePath) async {
     if (relativePath != null) {
@@ -29,20 +28,17 @@ class FileProcessing {
   // image 가져오기
   static Future<Map<String, dynamic>?> getFile() async {
     try {
-      const List<String> allowedExtensions = ['jpg', 'jpeg', 'png', 'pdf'];
       FilePickerResult? fileResult = await FilePicker.platform.pickFiles(
         type: FileType.custom,
         allowedExtensions: allowedExtensions,
         allowMultiple: false,
+        withData: true, // byte정보 포함
       );
 
       if (fileResult != null && fileResult.files.isNotEmpty) {
         String fileExtension = fileResult.files.first.extension!.toLowerCase();
-        File pickedFile = File(fileResult.files.first.path!);
-        Uint8List fileBytes = fileResult.files.single.bytes!;
-        Map<String, dynamic>? result =
-            await uploadFileToTmp(pickedFile, fileExtension, fileBytes);
-        return result;
+        Uint8List fileBytes = fileResult.files.first.bytes!;
+        return await uploadFileToTmp(fileExtension, fileBytes);
       } else {
         CustomToast.showToast('파일을 선택해주세요');
       }
@@ -60,13 +56,10 @@ class FileProcessing {
 
       if (imageFile != null) {
         // jpg, jpeg, png만 인식
-        var fileExtension = imageFile.path.split('.').last.toLowerCase();
+        String fileExtension = imageFile.path.split('.').last.toLowerCase();
         if (imageExtensions.contains(fileExtension)) {
-          File pickedFile = File(imageFile.path);
           Uint8List fileBytes = await imageFile.readAsBytes();
-          Map<String, dynamic>? result =
-              await uploadFileToTmp(pickedFile, fileExtension, fileBytes);
-          return result;
+          return await uploadFileToTmp(fileExtension, fileBytes);
         } else {
           // 이미지 파일이 아닌 경우 처리
           CustomToast.showToast('선택된 파일은 jpg, jpeg, png 파일이 아닙니다.');
@@ -82,21 +75,13 @@ class FileProcessing {
     return null;
   }
 
-  static Future<String?> getDownloadURL(String? relativePath) async {
-    if (relativePath != null) {
-      return await _storageRef.ref().child(relativePath).getDownloadURL();
-    } else {
-      return null;
-    }
-  }
-
   static Future<Map<String, dynamic>?> uploadFileToTmp(
-      File pickedFile, String fileExtension, Uint8List fileBytes) async {
+      String fileExtension, Uint8List fileBytes) async {
     try {
       String fileName = randomString(fileExtension);
       String relativePath = 'uploads/tmp/$fileName';
       var storageReference = _storageRef.ref().child(relativePath);
-      await storageReference.putFile(pickedFile);
+      await storageReference.putData(fileBytes);
       CustomToast.showToast('이미지 업로드 완료');
       return {
         'relativePath': relativePath,
@@ -111,50 +96,54 @@ class FileProcessing {
   }
 
   static Future<Map<String, String>?> transitionToStorage(
-      String relativePath, String fileName) async {
-    try {
-      String newRelativePath = 'uploads/saved/$fileName';
-      Reference sourceReference = _storageRef.ref(relativePath);
-      Reference destinationReference = _storageRef.ref(newRelativePath);
-      Uint8List? sourceData = await sourceReference.getData();
-      TaskSnapshot copyTask = await destinationReference.putData(sourceData!);
-      if (copyTask.state == TaskState.success) {
-        await deleteFile(relativePath);
-        print('파일 이동 성공');
-      } else {
-        print('파일 이동 실패: 복사 중 문제 발생');
+      String? relativePath, String? fileName, Uint8List? fileBytes) async {
+    if (relativePath != null) {
+      try {
+        String newRelativePath = 'uploads/saved/$fileName';
+        Reference destinationReference = _storageRef.ref(newRelativePath);
+        TaskSnapshot copyTask = await destinationReference.putData(fileBytes!);
+        if (copyTask.state == TaskState.success) {
+          await deleteFile(relativePath);
+          print('파일 이동 성공');
+        } else {
+          print('파일 이동 실패: 복사 중 문제 발생');
+        }
+        CustomToast.showToast('이미지 업로드 완료');
+        return {
+          'relativePath': newRelativePath,
+          'fileName': fileName!,
+        };
+      } catch (err) {
+        print('업로드 오류: $err');
+        CustomToast.showToast('이미지 업로드 오류');
       }
-      CustomToast.showToast('이미지 업로드 완료');
-      return {
-        'relativePath': newRelativePath,
-        'fileName': fileName,
-      };
-    } catch (err) {
-      print('업로드 오류: $err');
-      CustomToast.showToast('이미지 업로드 오류');
     }
     return null;
   }
 
-  static Future<void> deleteFile(String relativePath) async {
+  static Future<void> deleteFile(String? relativePath) async {
     // Firebase Storage 참조 얻기
-    try {
-      await _storageRef.ref().child(relativePath).delete();
-      print('업로드 된 파일이 성공적으로 삭제되었습니다.');
-      CustomToast.showToast('Photo delete 완료');
-    } catch (e, stackTrace) {
-      print('파일 삭제 중 오류 발생: $e\n$stackTrace');
-      CustomToast.showToast('Photo delete 에러 $e');
+    if (relativePath != null) {
+      try {
+        await _storageRef.ref().child(relativePath).delete();
+        print('업로드 된 파일이 성공적으로 삭제되었습니다.');
+        CustomToast.showToast('Photo delete 완료');
+      } catch (e, stackTrace) {
+        print('파일 삭제 중 오류 발생: $e\n$stackTrace');
+        CustomToast.showToast('Photo delete 에러 $e');
+      }
+    } else {
+      print('deleteFile: 삭제할 파일이 없음');
     }
   }
 
-  static Future<String?> fileToText(String? relativaPath) async {
-    if (relativaPath != null) {
+  static Future<String?> fileToText(String? relativePath) async {
+    if (relativePath != null) {
       try {
         final response = await http.post(
             Uri.parse('http://www.wooyoung-project.kro.kr/textTranslation'),
             headers: {'Content-Type': 'application/json'},
-            body: jsonEncode({'imageUrl': relativaPath}));
+            body: jsonEncode({'imageUrl': relativePath}));
         if (response.statusCode == 200) {
           String extractedText = response.body;
           print(extractedText);
